+++ lib/plex_assets.dart
@@ -1,3 +1,4 @@
 
 const String pdf = "packages/plex/assets/pdf.png";
 const String excel = "packages/plex/assets/excel.png";
+const String groupData = "packages/plex/assets/group_data.png";